django_prj/meiduo/meiduo_mall/requirements.txt
amqp==2.2.2
appdirs==1.4.3
billiard==3.5.0.3
celery==4.1.1
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.11
django-celery-results==1.0.1
django-ckeditor==5.6.1
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-daterange-filter==1.3.0
django-filter==2.0.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.0.1
django-js-asset==1.1.0
django-redis==4.8.0
django-redis-sessions==0.5.6
django-reversion==3.0.0
django-tinymce==2.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.3.0
drf-extensions==0.3.1
drf-haystack==1.8.4
elasticsearch==6.3.1
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.16.0
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
jieba==0.38
Jinja2==2.10
kombu==4.2.0
MarkupSafe==1.0
mutagen==1.38
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.7
packaging==16.8
Pillow==4.3.0
pkg-resources==0.0.0
pycryptodomex==3.6.1
PyJWT==1.6.3
pymongo==3.4.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-alipay-sdk==1.8.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
QQLoginTool==0.3.0
redis==2.10.5
redis-py-cluster==1.3.4
requests==2.18.4
reversion==0.2
serializers==0.2.4
six==1.10.0
tablib==0.12.1
toml==0.9.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
Whoosh==2.7.4
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
